GameLog-API/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.21.1
pydantic[email]==1.10.7
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
